# docs/swagger.yaml
definitions:
...
  auth.UserRegister:
    properties:
      confirm_password:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  model.Broadcast:
    properties:
      content:
        type: string
      create_time:
        type: string
      id:
        type: integer
      path:
        type: string
      picture:
        type: string
      sha:
        type: string
      title:
        type: string
    type: object
  model.Place:
    properties:
      area:
        description: 地点面积
        type: string
      data:
        description: 地点资料
        type: string
      entertainment:
        description: 娱乐项目数量
        type: string
      id:
        type: integer
      name:
        description: 地点名称
        type: string
      path1:
        type: string
      path2:
        type: string
      path3:
        type: string
      picture1:
        description: 地点轮换封面1
        type: string
      picture2:
        description: 地点轮换封面2
        type: string
      picture3:
        description: 地点轮换封面3
        type: string
      scenic_spot:
        description: 特色景点数量
        type: string
      sha1:
        type: string
      sha2:
        type: string
      sha3:
        type: string
      visitor:
        description: 最大游客量
        type: string
    type: object
  model.Script:
    properties:
      avatar:
        description: 剧本封面
        type: string
      brief_intro:
        description: 剧本简介
        type: string
      id:
        type: integer
      introduction:
        description: 剧本介绍
        type: string
      path:
        type: string
      place:
        description: 地点
        type: string
      price:
        description: 价格
        type: integer
      script_name:
        description: 剧本名称
        type: string
      sha:
        type: string
      tag1:
        description: 标签一
        type: string
      tag2:
        description: 标签二
        type: string
      tag3:
        description: 标签三
        type: string
      tag4:
        description: 标签四
        type: string
      tag5:
        description: 标签五
        type: string
      time:
        description: 剧本时长
        type: string
    type: object
...
  place.PlaceName:
    properties:
      place_name:
        description: 地点名称
        type: string
    type: object
  place.Places:
    properties:
      area:
        description: 地点面积
        type: string
      data:
        description: 地点资料
        type: string
      entertainment:
        description: 娱乐项目数量
        type: string
      name:
        description: 地点名称
        type: string
      scenic_spot:
        description: 特色景点数量
        type: string
      visitor:
        description: 最大游客量
        type: string
    type: object
  script.ScriptEditInfor:
    properties:
      brief_intro:
        description: 剧本简介
        type: string
      id:
        type: integer
      introduction:
        description: 剧本介绍
        type: string
      place:
        description: 地点
        type: string
      price:
        description: 价格
        type: integer
      script_name:
        description: 剧本名称
        type: string
      tag1:
        description: 标签一
        type: string
      tag2:
        description: 标签二
        type: string
      tag3:
        description: 标签三
        type: string
      tag4:
        description: 标签四
        type: string
      tag5:
        description: 标签五
        type: string
      time:
        description: 剧本时间
        type: string
    type: object
...
  script.Scripts:
    properties:
      brief_intro:
        description: 剧本简介
        type: string
      introduction:
        description: 剧本介绍
        type: string
      place:
        description: 地点
        type: string
      price:
        description: 价格
        type: integer
      script_name:
        description: 剧本名称
        type: string
      tag1:
        description: 标签一
        type: string
      tag2:
        description: 标签二
        type: string
      tag3:
        description: 标签三
        type: string
      tag4:
        description: 标签四
        type: string
      tag5:
        description: 标签五
        type: string
      time:
        description: 剧本时间
        type: string
    type: object
...
host: localhost:8918
info:
  contact:
    email: 2295616516@qq.com
    name: Eternal-Faith
  description: 山乡剧变
  termsOfService: http://swagger.io/terrms
  title: uima
  version: 1.0.0
paths:
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: '"注册一个新用户"'
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.UserRegister'
      produces:
      - application/json
      responses:
        "200":
          description: 用户创建成功
        "400":
          description: 输入有误，格式错误
        "401":
          description: 电话号码重复
      summary: '"注册"'
      tags:
      - auth
  /broadcast/:broadcast_id:
    delete:
      consumes:
      - application/json
      description: '"删除一个精彩放送的信息"'
      parameters:
      - description: broadcast_id
        in: path
        name: broadcast_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
        "400":
          description: 删除失败
      summary: '"删除一个精彩放送信息"'
      tags:
      - bro
  /broadcast/avatar/:broadcast_id:
    post:
      consumes:
      - application/json
      description: '"精彩放送图片"'
      parameters:
      - description: broadcast_id
        in: path
        name: broadcast_id
        required: true
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
        "400":
          description: 上传失败
        "401":
          description: 身份验证失败
      summary: '"修改精彩放送图片"'
      tags:
      - bro
...
  /broadcast/get_all:
    get:
      consumes:
      - application/json
      description: '"精彩放送"'
      parameters:
      - description: Broadcast
        in: body
        name: broadcast
        required: true
        schema:
          $ref: '#/definitions/model.Broadcast'
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
        "400":
          description: 上传失败
        "401":
          description: 身份验证失败
      summary: '"精彩放送"'
      tags:
      - bro
...
  /place/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 地点
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/place.Places'
      responses:
        "200":
          description: 创建成功
        "500":
          description: 服务器错误
      summary: '"创建地点信息"'
      tags:
      - place
...
  /place/picturethree:
    post:
      consumes:
      - application/json
      description: '"修改地点图片2"'
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: id--地点的id
        in: formData
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"mgs":"success"}'
          schema:
            $ref: '#/definitions/model.Script'
        "400":
          description: 上传失败,请检查token与其他配置参数是否正确
      summary: '"修改地点图片2"'
      tags:
      - place
  /place/picturetwo:
    post:
      consumes:
      - application/json
      description: '"修改地点图片3"'
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: id--地点的id
        in: formData
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"mgs":"success"}'
          schema:
            $ref: '#/definitions/model.Script'
        "400":
          description: 上传失败,请检查token与其他配置参数是否正确
      summary: '"修改地点图片3"'
      tags:
      - place
  /place/view:
    post:
      consumes:
      - application/json
      description: '"查看地点信息"'
      parameters:
      - description: 地点名称
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/place.PlaceName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Place'
        "500":
          description: 服务器错误
      summary: '"查看地点"'
      tags:
      - place
...
  /script/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 剧本
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/script.Scripts'
      responses:
        "200":
          description: 创建成功
        "400":
          description: 输入格式有误
        "500":
          description: 服务器错误
      summary: '"创建剧本"'
      tags:
      - script
  /script/edit:
    post:
      consumes:
      - application/json
      description: '"修改剧本内容"'
      parameters:
      - description: 剧本
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/script.ScriptEditInfor'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
        "400":
          description: 输入格式有误
        "500":
          description: 服务器错误
      summary: '"修改剧本信息"'
      tags:
      - script
  /script/interface:
    get:
      consumes:
      - application/json
      description: '"返回所有剧本的封面、名字、ID"'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            items:
              $ref: '#/definitions/model.Script'
            type: array
        "404":
          description: 获取失败
      summary: '"剧本主界面"'
      tags:
      - script
...
schemes:
- http
swagger: "2.0"